refactor: use public functions in scraper module instead of controller directly
