+++ worldometer/scraper/__init__.py
@@ -1,0 +1,9 @@
+__all__ = [
+    'get_data_tables',
+    'get_rts_counters_object'
+]
+
+from worldometer.scraper.controller import (
+    get_data_tables,
+    get_rts_counters_object
+)
